requirements.txt
MetaTrader5>=5.0.45
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
tensorboard>=2.13.0
gymnasium>=0.29.0
requests>=2.31.0
colorama>=0.4.6
